test: divide tests into core and heavy
