+++ .github/workflows/core-tests.yaml
@@ -1,12 +1,13 @@
-name: Tests
+name: Core tests
 
 on:
   pull_request:
-    branches: [ dev, main ]
+    branches: [ main ]
     paths:
       - 'datadreamer/**/**.py'
-      - 'tests/**/**.py'
-      - .github/workflows/tests.yaml
+      - 'tests/core_tests/**/**.py'
+      - .github/workflows/core-tests.yaml
+  workflow_dispatch:
 
 jobs:
   run_tests:
@@ -14,7 +15,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [buildjet-8vcpu-ubuntu-2204, windows-latest, macOS-latest]
-        version: ['3.10']
+        version: ['3.10', '3.11']
 
     runs-on: ${{ matrix.os }}
 
@@ -49,11 +50,11 @@ jobs:
         pip install coverage-badge>=1.1.0 pytest-cov>=4.1.0
     - name: Run tests with coverage [Ubuntu]
       if: matrix.os == 'buildjet-8vcpu-ubuntu-2204' && matrix.version == '3.10'
-      run: pytest tests --cov=datadreamer --cov-report xml --junit-xml pytest.xml
+      run: pytest tests/core_tests --cov=datadreamer --cov-report xml --junit-xml pytest.xml
 
     - name: Run tests [Windows, macOS]
-      if: matrix.os != 'buildjet-8vcpu-ubuntu-2204' || matrix.version != '3.10'
-      run: pytest tests --junit-xml pytest.xml
+      if: matrix.os != 'buildjet-8vcpu-ubuntu-2204'
+      run: pytest tests/core_tests --junit-xml pytest.xml
 
     - name: Generate coverage badge [Ubuntu]
       if: matrix.os == 'buildjet-8vcpu-ubuntu-2204' && matrix.version == '3.10'
@@ -93,7 +94,7 @@ jobs:
   publish-test-results:
     name: "Publish Tests Results"
     needs: run_tests
-    runs-on: ubuntu-latest
+    runs-on: buildjet-8vcpu-ubuntu-2204
     permissions:
       checks: write
       pull-requests: write
