+++ tests/heavy_tests/integration/test_pipeline.py
@@ -8,10 +8,9 @@ import pytest
 import torch
 
 # Get the total memory in GB
-total_memory = psutil.virtual_memory().total / (1000**3)
-total_memory = int(total_memory) + (total_memory > int(total_memory))
+total_memory = psutil.virtual_memory().total / (1024*3)
 # Get the total disk space in GB
-total_disk_space = psutil.disk_usage("/").total / (1000**3)
+total_disk_space = psutil.disk_usage("/").total / (1024**3)
 
 
 def _check_detection_pipeline(cmd: str, target_folder: str):
@@ -45,220 +44,12 @@ def _check_detection_pipeline(cmd: str, target_folder: str):
     ), "bboxes_visualization directory not created"
 
 
-def _check_wrong_argument_choice(cmd: str):
-    with pytest.raises(subprocess.CalledProcessError):
-        subprocess.check_call(cmd, shell=True)
-
-
-def _check_wrong_value(cmd: str):
-    with pytest.raises(ValueError):
-        try:
-            subprocess.check_output(cmd, shell=True, stderr=subprocess.STDOUT)
-        except subprocess.CalledProcessError as e:
-            raise ValueError(e.output.decode()) from e
-
-
-# =========================================================
-# ARGUMENTS CHECKS
-# =========================================================
-def test_invalid_task_value():
-    # Define the cmd
-    cmd = "datadreamer --task invalid_task"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_prompts_number_type():
-    # Define the cmd
-    cmd = "datadreamer --prompts_number value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_num_objects_range_type():
-    # Define the cmd
-    cmd = "datadreamer --num_objects_range value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_conf_threshold_range_type():
-    # Define the cmd
-    cmd = "datadreamer --conf_threshold value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_image_tester_patience_type():
-    # Define the cmd
-    cmd = "datadreamer --image_tester_patience value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_seed_type():
-    # Define the cmd
-    cmd = "datadreamer --seed value --device cpu"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_prompt_generator():
-    # Define the cmd
-    cmd = "datadreamer --prompt_generator invalide_value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_image_generator():
-    # Define the cmd
-    cmd = "datadreamer --image_generator invalide_value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_image_annotator():
-    # Define the cmd
-    cmd = "datadreamer --image_annotator invalide_value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_det_image_annotator():
-    # Define the cmd
-    cmd = "datadreamer --image_annotator clip"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_clf_image_annotator():
-    # Define the cmd
-    cmd = "datadreamer --image_annotator owlv2 --task classification"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_device():
-    # Define the cmd
-    cmd = "datadreamer --device invalide_value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_invalid_annotator_size():
-    # Define the cmd
-    cmd = "datadreamer --annotator_size invalide_value"
-    _check_wrong_argument_choice(cmd)
-
-
-def test_empty_class_names():
-    # Define the cmd
-    cmd = "datadreamer --class_names []"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_class_names():
-    # Define the cmd
-    cmd = "datadreamer --class_names [2, -1]"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_prompts_number():
-    # Define the cmd
-    cmd = "datadreamer --prompts_number -1"
-    _check_wrong_value(cmd)
-
-
-def test_negative_conf_threshold():
-    # Define the cmd
-    cmd = "datadreamer --conf_threshold -1"
-    _check_wrong_value(cmd)
-
-
-def test_big_conf_threshold():
-    # Define the cmd
-    cmd = "datadreamer --conf_threshold 10"
-    _check_wrong_value(cmd)
-
-
-def test_negative_annotation_iou_threshold():
-    # Define the cmd
-    cmd = "datadreamer --annotation_iou_threshold -1"
-    _check_wrong_value(cmd)
-
-
-def test_big_annotation_iou_threshold():
-    # Define the cmd
-    cmd = "datadreamer --annotation_iou_threshold 10"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_image_tester_patience():
-    # Define the cmd
-    cmd = "datadreamer --image_tester_patience -1"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_seed():
-    # Define the cmd
-    cmd = "datadreamer --seed -1 --device cpu"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_synonym_generator():
-    # Define the cmd
-    cmd = "datadreamer --device cpu --synonym_generator invalid"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_lm_quantization():
-    # Define the cmd
-    cmd = "datadreamer --device cude --lm_quantization invalid"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_device_lm_quantization():
-    # Define the cmd
-    cmd = "datadreamer --device cpu --lm_quantization 4bit"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_batch_size_prompt():
-    # Define the cmd
-    cmd = "datadreamer --batch_size_prompt -1"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_batch_size_annotation():
-    # Define the cmd
-    cmd = "datadreamer --batch_size_annotation -1"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_batch_size_image():
-    # Define the cmd
-    cmd = "datadreamer --batch_size_image -1"
-    _check_wrong_value(cmd)
-
-
-def test_invalid_num_objects_range():
-    # Define the cmd
-    cmd = "datadreamer --num_objects_range 1"
-    _check_wrong_value(cmd)
-
-
-def test_many_num_objects_range():
-    # Define the cmd
-    cmd = "datadreamer --num_objects_range 1 2 3"
-    _check_wrong_value(cmd)
-
-
-def test_desc_num_objects_range():
-    # Define the cmd
-    cmd = "datadreamer --num_objects_range 3 1"
-    _check_wrong_value(cmd)
-
-
-def test_negative_num_objects_range():
-    # Define the cmd
-    cmd = "datadreamer --num_objects_range -3 1"
-    _check_wrong_value(cmd)
-
-
 # =========================================================
 # DETECTION - SIMPLE LM
 # =========================================================
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_simple_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -279,8 +70,8 @@ def test_cpu_simple_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -301,8 +92,8 @@ def test_cuda_simple_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 55,
-    reason="Test requires GPU, at least 15GB of RAM and 55GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 55,
+    reason="Test requires GPU, at least 16GB of RAM and 55GB of HDD",
 )
 def test_cuda_simple_llm_synonym_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -324,8 +115,8 @@ def test_cuda_simple_llm_synonym_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_wordnet_synonym_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -347,8 +138,8 @@ def test_cuda_simple_wordnet_synonym_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_simple_sdxl_detection_pipeline():
     # Define target folder
@@ -369,8 +160,8 @@ def test_cpu_simple_sdxl_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_sdxl_detection_pipeline():
     # Define target folder
@@ -391,8 +182,8 @@ def test_cuda_simple_sdxl_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_simple_sdxl_lightning_detection_pipeline():
     # Define target folder
@@ -413,8 +204,8 @@ def test_cpu_simple_sdxl_lightning_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_sdxl_lightning_detection_pipeline():
     # Define target folder
@@ -460,8 +251,8 @@ def test_cpu_lm_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 55,
-    reason="Test requires at least 15GB of RAM, CUDA support and 55GB of HDD",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 55,
+    reason="Test requires at least 16GB of RAM, CUDA support and 55GB of HDD",
 )
 def test_cuda_lm_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -527,8 +318,8 @@ def test_cpu_lm_sdxl_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 55,
-    reason="Test requires at least 15GB of RAM, CUDA support and 55GB of HDD",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 55,
+    reason="Test requires at least 16GB of RAM, CUDA support and 55GB of HDD",
 )
 def test_cuda_lm_sdxl_detection_pipeline():
     # Define target folder
@@ -575,8 +366,8 @@ def test_cuda_4bit_lm_sdxl_detection_pipeline():
 # DETECTION - TinyLlama LLM
 # =========================================================
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_tiny_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -597,8 +388,8 @@ def test_cpu_tiny_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_tiny_sdxl_turbo_detection_pipeline():
     # Define target folder
@@ -619,8 +410,8 @@ def test_cuda_tiny_sdxl_turbo_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_tiny_sdxl_detection_pipeline():
     # Define target folder
@@ -641,8 +432,8 @@ def test_cpu_tiny_sdxl_detection_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_tiny_sdxl_detection_pipeline():
     # Define target folder
@@ -666,8 +457,8 @@ def test_cuda_tiny_sdxl_detection_pipeline():
 # CLASSIFICATION - SIMPLE LM
 # =========================================================
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_simple_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -690,8 +481,8 @@ def test_cpu_simple_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -714,8 +505,8 @@ def test_cuda_simple_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 55,
-    reason="Test requires GPU, at least 15GB of RAM and 55GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 55,
+    reason="Test requires GPU, at least 16GB of RAM and 55GB of HDD",
 )
 def test_cuda_simple_llm_synonym_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -739,8 +530,8 @@ def test_cuda_simple_llm_synonym_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_wordnet_synonym_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -764,8 +555,8 @@ def test_cuda_simple_wordnet_synonym_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_simple_sdxl_classification_pipeline():
     # Define target folder
@@ -788,8 +579,8 @@ def test_cpu_simple_sdxl_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_simple_sdxl_classification_pipeline():
     # Define target folder
@@ -839,8 +630,8 @@ def test_cpu_lm_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 55,
-    reason="Test requires at least 15GB of RAM, 55GB of HDD and CUDA support",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 55,
+    reason="Test requires at least 16GB of RAM, 55GB of HDD and CUDA support",
 )
 def test_cuda_lm_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -912,8 +703,8 @@ def test_cpu_lm_sdxl_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 55,
-    reason="Test requires at least 15GB of RAM, CUDA support and 55GB of HDD",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 55,
+    reason="Test requires at least 16GB of RAM, CUDA support and 55GB of HDD",
 )
 def test_cuda_lm_sdxl_classification_pipeline():
     # Define target folder
@@ -964,8 +755,8 @@ def test_cuda_4bit_lm_sdxl_classification_pipeline():
 # CLASSIFICATION - TinyLlama LLM
 # =========================================================
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_tiny_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -988,8 +779,8 @@ def test_cpu_tiny_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_tiny_sdxl_turbo_classification_pipeline():
     # Define target folder
@@ -1012,8 +803,8 @@ def test_cuda_tiny_sdxl_turbo_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
+    total_memory < 16 or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM and 35GB of HDD",
 )
 def test_cpu_tiny_sdxl_classification_pipeline():
     # Define target folder
@@ -1036,8 +827,8 @@ def test_cpu_tiny_sdxl_classification_pipeline():
 
 
 @pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
+    not torch.cuda.is_available() or total_memory < 16 or total_disk_space < 35,
+    reason="Test requires GPU, at least 16GB of RAM and 35GB of HDD",
 )
 def test_cuda_tiny_sdxl_classification_pipeline():
     # Define target folder
@@ -1053,85 +844,6 @@ def test_cuda_tiny_sdxl_classification_pipeline():
         f"--image_annotator clip "
         f"--image_generator sdxl "
         f"--use_image_tester "
-        f"--device cuda"
-    )
-    # Check the run of the pipeline
-    _check_detection_pipeline(cmd, target_folder)
-
-
-# =========================================================
-# TEST WITH CONFIG FILE
-# =========================================================
-@pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
-)
-def test_cpu_simple_sdxl_turbo_config_detection_pipeline():
-    # Define target folder
-    target_folder = "data/data-det-cpu-simple-sdxl-turbo-config/"
-    # Define the command to run the datadreamer
-    cmd = (
-        f"datadreamer --save_dir {target_folder} "
-        f"--num_objects_range 1 2 "
-        f"--config ./sample_config.yaml "
-        f"--device cpu"
-    )
-    # Check the run of the pipeline
-    _check_detection_pipeline(cmd, target_folder)
-
-
-@pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
-)
-def test_cuda_simple_sdxl_turbo_config_detection_pipeline():
-    # Define target folder
-    target_folder = "data/data-det-cuda-simple-sdxl-turbo-config/"
-    # Define the command to run the datadreamer
-    cmd = (
-        f"datadreamer --save_dir {target_folder} "
-        f"--num_objects_range 1 2 "
-        f"--config ./sample_config.yaml "
-        f"--device cuda"
-    )
-    # Check the run of the pipeline
-    _check_detection_pipeline(cmd, target_folder)
-
-
-@pytest.mark.skipif(
-    total_memory < 15 or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM and 35GB of HDD",
-)
-def test_cpu_simple_sdxl_turbo_config_classification_pipeline():
-    # Define target folder
-    target_folder = "data/data-cls-cpu-simple-sdxl-turbo-config/"
-    # Define the command to run the datadreamer
-    cmd = (
-        f"datadreamer --task classification "
-        f"--save_dir {target_folder} "
-        f"--num_objects_range 1 2 "
-        f"--image_annotator clip "
-        f"--config ./sample_config.yaml "
-        f"--device cpu"
-    )
-    # Check the run of the pipeline
-    _check_detection_pipeline(cmd, target_folder)
-
-
-@pytest.mark.skipif(
-    not torch.cuda.is_available() or total_memory < 15 or total_disk_space < 35,
-    reason="Test requires GPU, at least 15GB of RAM and 35GB of HDD",
-)
-def test_cuda_simple_sdxl_turbo_config_classification_pipeline():
-    # Define target folder
-    target_folder = "data/data-cls-cuda-simple-sdxl-turbo-config/"
-    # Define the command to run the datadreamer
-    cmd = (
-        f"datadreamer --task classification "
-        f"--save_dir {target_folder} "
-        f"--num_objects_range 1 2 "
-        f"--image_annotator clip "
-        f"--config ./sample_config.yaml "
         f"--device cuda"
     )
     # Check the run of the pipeline
