+++ tests/heavy_tests/unittests/test_prompt_generation.py
@@ -15,31 +15,9 @@ from datadreamer.prompt_generation.wordnet_synonym_generator import (
 )
 
 # Get the total memory in GB
-total_memory = psutil.virtual_memory().total / (1000**3)
-total_memory = int(total_memory) + (total_memory > int(total_memory))
+total_memory = psutil.virtual_memory().total / (1024**3)
 # Get the total disk space in GB
-total_disk_space = psutil.disk_usage("/").total / (1000**3)
-
-
-def test_simple_prompt_generator():
-    class_names = ["dog", "cat", "bird", "tree", "car", "person", "house", "flower"]
-    prompt_generator = SimplePromptGenerator(class_names, prompts_number=10)
-    prompts = prompt_generator.generate_prompts()
-    # Check that the some prompts were generated
-    assert len(prompts) > 0
-    # Iterate through the prompts
-    for selected_objects, prompt_text in prompts:
-        # Selected objects aren't empty
-        assert len(selected_objects) > 0
-        # The slected objects are in the range
-        assert (
-            prompt_generator.num_objects_range[0]
-            <= len(selected_objects)
-            <= prompt_generator.num_objects_range[1]
-        )
-        # Check the generated text
-        assert prompt_text == f"A photo of a {', a '.join(selected_objects)}"
-
+total_disk_space = psutil.disk_usage("/").total / (1024**3)
 
 def _check_lm_prompt_generator(
     device: str, prompt_generator_class=LMPromptGenerator, quantization: str = "none"
@@ -70,20 +48,11 @@ def _check_lm_prompt_generator(
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM, 35GB of HDD and CUDA support",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM, 35GB of HDD and CUDA support",
 )
 def test_cuda_lm_prompt_generator():
     _check_lm_prompt_generator("cuda")
-
-
-@pytest.mark.skipif(
-    total_memory < 12 or not torch.cuda.is_available() or total_disk_space < 25,
-    reason="Test requires at least 12GB of RAM, 25GB of HDD and CUDA support",
-)
-def test_cuda_4bit_lm_prompt_generator():
-    _check_lm_prompt_generator("cuda", quantization="4bit")
-
 
 @pytest.mark.skipif(
     total_memory < 32 or total_disk_space < 35,
@@ -91,23 +60,6 @@ def test_cuda_4bit_lm_prompt_generator():
 )
 def test_cpu_lm_prompt_generator():
     _check_lm_prompt_generator("cpu")
-
-
-@pytest.mark.skipif(
-    total_memory < 8 or not torch.cuda.is_available() or total_disk_space < 12,
-    reason="Test requires at least 8GB of RAM, 12GB of HDD and CUDA support",
-)
-def test_cuda_tinyllama_lm_prompt_generator():
-    _check_lm_prompt_generator("cuda", TinyLlamaLMPromptGenerator)
-
-
-@pytest.mark.skipif(
-    total_memory < 12 or total_disk_space < 12,
-    reason="Test requires at least 12GB of RAM and 12GB of HDD for running on CPU",
-)
-def test_cpu_tinyllama_lm_prompt_generator():
-    _check_lm_prompt_generator("cpu", TinyLlamaLMPromptGenerator)
-
 
 def _check_synonym_generator(device: str, synonym_generator_class=LMSynonymGenerator):
     synonyms_num = 3
@@ -128,8 +80,8 @@ def _check_synonym_generator(device: str, synonym_generator_class=LMSynonymGener
 
 
 @pytest.mark.skipif(
-    total_memory < 15 or not torch.cuda.is_available() or total_disk_space < 35,
-    reason="Test requires at least 15GB of RAM, 35GB of HDD and CUDA support",
+    total_memory < 16 or not torch.cuda.is_available() or total_disk_space < 35,
+    reason="Test requires at least 16GB of RAM, 35GB of HDD and CUDA support",
 )
 def test_cuda_synonym_generator():
     _check_synonym_generator("cuda")
@@ -141,15 +93,3 @@ def test_cuda_synonym_generator():
 )
 def test_cpu_synonym_generator():
     _check_synonym_generator("cpu")
-
-
-def test_cpu_wordnet_synonym_generator():
-    _check_synonym_generator("cpu", WordNetSynonymGenerator)
-
-
-@pytest.mark.skipif(
-    torch.cuda.is_available(),
-    reason="Test requires CUDA support",
-)
-def test_cuda_wordnet_synonym_generator():
-    _check_synonym_generator("cuda", WordNetSynonymGenerator)
